Bug: activeNav update
- adjusted nav li's to also have their corresponding section as their 
class to allow activeNav to work, used to assign active for nav whose 
section is in viewport
- bc of for loop changes, need to adjust what is passed to activeNav
TODO: Why don't activeNav() and scrollToSxn() work when they are both 
on.  Turned on independently they both work. But when they are together, 
scrolltoSxn does not scroll and selected tabs don't become inactive when 
we scroll to a different section.